requirements.txt
discord.py
python-dotenv
asyncio
pillow
requests